0.4.2
- **Updated**
  - `Reaction Tracker`
    - Fixed error of trying to delete a reaction used that doesn't exist

+++ scripts/lib/reactionTracker.js
@@ -46,7 +46,8 @@ async function removeEncounterReactions(encounter) {
         const existing = actor.itemTypes.effect.find(
             (e) => e.slug === "effect-reaction-used"
         );
-        existing.delete();
+        if (existing)
+            existing.delete();
     }
 }
